+++ NEWS.md
@@ -1,11 +1,13 @@
 * update data
 * Update team group
+* fix link to linkedin account
 
 # covid19ita 0.16.0
 
 * Update data
 * Added focus section for Vercelli and Alessandria
 * Update istruzioni (ita)
+* update team group eng
 * added alessandria and vercelli
 * translated x-y labels of the graphs in the epidemic section (#4)
 * translated Novara 2020-03-20 (#20)